hotfix - insert data in rds
